+Unsubscribe to current restaurant when subscribe to new; +rcsSocket handle restaurant message event; +Allow dup-login of the same user; +Fix client logout logic

+++ assets/linker/js/client/service/authService.js
@@ -13,11 +13,16 @@ angular
           });
       };
 
-      authService.logout = function () {
+      authService.logout = function (cb) {
+        if (!authService.isAuthenticated()) {
+          return cb();
+        }
+
         return rcsAPI.User.logout()
           .success(function () {
             SessionService.destroy();
             $rootScope.$broadcast(AUTH_EVENTS.logoutSuccess);
+            cb();
           });
       };
